moved components into components folder and set up react-router-dom
