+++ client/components/Footer.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export const Footer = () => {
+  return <div />;
+};
