+++ client/components/Nav.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export const Nav = () => {
+  return <div />;
+};
