+++ client/components/index.js
@@ -1,0 +1,4 @@
+export * from './Challenges';
+export * from './IndividualChallenge';
+export * from './Footer';
+export * from './Nav';
