+++ client/index.js
@@ -1,7 +1,13 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import { HashRouter } from 'react-router-dom';
 import App from './App';
 
 const root = document.querySelector('#root');
 
-ReactDOM.render(<App />, root);
+ReactDOM.render(
+  <HashRouter>
+    <App />
+  </HashRouter>,
+  root,
+);
